Add custom exceptions to VictimController

+++ src/main/java/se/iths/crimedatabase/controller/VictimController.java
@@ -4,7 +4,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import se.iths.crimedatabase.entity.Victim;
+import se.iths.crimedatabase.exception.BadRequestException;
 import se.iths.crimedatabase.exception.MethodNotAllowedException;
+import se.iths.crimedatabase.exception.NotFoundException;
 import se.iths.crimedatabase.service.VictimService;
 
 import java.util.Optional;
@@ -20,18 +22,27 @@ public class VictimController {
 
     @PostMapping
     public ResponseEntity<Victim> create(@RequestBody Victim victim) {
+        if (victim.getFirstName().isEmpty() || victim.getLastName().isEmpty())
+            throw new BadRequestException("First and last name cannot be empty");
+
         Victim createdVictim = victimService.create(victim);
         return new ResponseEntity<>(createdVictim, HttpStatus.CREATED);
     }
 
     @DeleteMapping("{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (victimService.findById(id).isEmpty())
+            throw new NotFoundException(notFoundMessage(id));
+
         victimService.delete(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Optional<Victim>> findById(@PathVariable Long id) {
+        if (victimService.findById(id).isEmpty())
+            throw new NotFoundException(notFoundMessage(id));
+
         Optional<Victim> foundVictim = victimService.findById(id);
         return new ResponseEntity<>(foundVictim, HttpStatus.OK);
     }
@@ -39,6 +50,9 @@ public class VictimController {
     @GetMapping
     public ResponseEntity<Iterable<Victim>> findAll() {
         Iterable<Victim> victims = victimService.findAll();
+        if (!victims.iterator().hasNext())
+            throw new NotFoundException("No victims found");
+
         return new ResponseEntity<>(victims, HttpStatus.OK);
     }
